+++ test/rocprim/test_device_merge_sort.cpp
@@ -66,8 +66,7 @@ typedef ::testing::Types<
     DeviceSortParams<int8_t, int8_t>,
     DeviceSortParams<uint8_t, uint8_t>,
     DeviceSortParams<rocprim::half, rocprim::half, test_utils::half_less>,
-    //TODO: Disable bfloat16 test until we get a better bfloat16 implemetation for host side
-    //DeviceSortParams<rocprim::bfloat16, rocprim::bfloat16, test_utils::bfloat16_less>,
+    DeviceSortParams<rocprim::bfloat16, rocprim::bfloat16, test_utils::bfloat16_less>,
     DeviceSortParams<int, float, ::rocprim::greater<int>>,
     DeviceSortParams<short, test_utils::custom_test_type<int>>,
     DeviceSortParams<double, test_utils::custom_test_type<double>>,
@@ -244,7 +243,7 @@ TYPED_TEST(RocprimDeviceSortTests, SortKeyValue)
             std::vector<key_type> keys_input = test_utils::get_random_data<key_type>(size, -100, 100, seed_value); // float16 can't exceed 65504
 
             std::vector<value_type> values_input(size);
-            std::iota(values_input.begin(), values_input.end(), 0);
+            test_utils::iota(values_input.begin(), values_input.end(), 0);
 
             std::vector<key_type> keys_output(size);
             std::vector<value_type> values_output(size);