+++ test/rocprim/test_device_select.cpp
@@ -61,7 +61,7 @@ typedef ::testing::Types<
     DeviceSelectParams<int8_t, int8_t>,
     DeviceSelectParams<uint8_t, uint8_t>,
     DeviceSelectParams<rocprim::half, rocprim::half>,
-    //TODO: Disable bfloat16 test until we get a better bfloat16 implemetation for host side
+    //TODO: Disable bfloat16 test until the follwing PR merge: https://github.com/ROCm-Developer-Tools/HIP/pull/2303
     //DeviceSelectParams<rocprim::bfloat16, rocprim::bfloat16>,
     DeviceSelectParams<unsigned char, float, int, true>,
     DeviceSelectParams<double, double, int, true>,