+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Full documentation for rocPRIM is available at [https://codedocs.xyz/ROCmSoftwar
 ## [Unreleased rocPRIM-2.10.12 for ROCm 5.0.0]
 ### Fixed
 - Enable bfloat16 tests and reduce threshold for bfloat16
+- Fix device scan limit_size feature
 ### Addded
 - Added scan size limit feature
 - Add block_load_striped and block_store_striped
@@ -14,6 +15,7 @@ Full documentation for rocPRIM is available at [https://codedocs.xyz/ROCmSoftwar
 ### Known issues
 - Unit tests may soft hang on MI200 when running in hipMallocManaged mode.
 - block_histogram, device_scan unit tests failing for HIP on Windows
+- ReduceEmptyInput cause random faulire with bfloat16
 
 ## [Unreleased rocPRIM-2.10.11 for ROCm 4.5.0]
 ### Addded